fix big integer foreign key

+++ database/migrations/2019_09_17_081925_create_bands_table.php
@@ -18,7 +18,7 @@ class CreateBandsTable extends Migration
             $table->integer("min");
             $table->integer("max");
             $table->string("color");
-            $table->integer("antennas_id");
+            $table->bigInteger("antennas_id");
             $table->foreign('antennas_id')->references('id')->on('antennas');
         });
     }
